--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,12 +1,15 @@
 package hw10programoptimization
 
 import (
+	"bufio"
 	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"regexp"
 	"strings"
+
+	jsoniter "github.com/json-iterator/go"
 )
 
 type User struct {
@@ -22,6 +25,36 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
+	domain = "." + domain
+	scanner := bufio.NewScanner(r)
+
+	result := make(DomainStat)
+
+	json := jsoniter.ConfigCompatibleWithStandardLibrary
+	var user User
+	for scanner.Scan() {
+		bytes := scanner.Bytes()
+
+		if strings.Contains(string(bytes), domain) {
+			user = User{}
+			if err := json.Unmarshal(bytes, &user); err != nil {
+				return DomainStat{}, err
+			}
+
+			matched := strings.HasSuffix(user.Email, domain)
+
+			if matched {
+				num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
+				num++
+				result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+			}
+		}
+	}
+
+	return result, nil
+}
+
+func GetDomainStatSlow(r io.Reader, domain string) (DomainStat, error) {
 	u, err := getUsers(r)
 	if err != nil {
 		return nil, fmt.Errorf("get users error: %w", err)